book_scraper/requirements.txt
scrapy>=2.11.0
mysql-connector-python>=8.0.0
pymongo>=4.0.0
